feat(mgmt): reorg config building
Add a new submodule to centralize the building of configurations.
The submodule contains the logic to:
 - build internal config from external
 - build router config from the internal config and the state
   created by the vpc-manager.

The frr config is a partial rendering of the internal config.
The vpc manager config resides in the vpc-manager.
