+++ mgmt/src/processor/confbuild/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Open Network Fabric Authors
+
+//! Submodule to build distinct configurations
+
+pub(crate) mod internal;
+pub(crate) mod namegen;
+pub(crate) mod router;
