+++ mgmt/src/processor/confbuild/namegen.rs
