+++ mgmt/src/processor/mod.rs
@@ -8,6 +8,5 @@ pub mod confbuild;
 mod display;
 pub mod gwconfigdb;
 pub mod launch;
-mod namegen;
 pub mod proc;
 pub mod tests;
